change copyright years, and rename jdk.internal.util.Hex256 to jdk.internal.digits.Hex256, and add comments.
